+++ server/db/FUN_PRO_VIEWS/15_proc_get_dataTable_with_filter.sql
@@ -25,7 +25,7 @@ BEGIN
     ) INTO @sql
     FROM  dataTables t
 	INNER JOIN dataCols col on (t.id = col.id_dataTables)
-    WHERE t.id = table_id_param AND col.`name` NOT IN ('id_users','record_id','user_name','id_actionTriggerTypes','triggerType', 'entry_date');
+    WHERE t.id = table_id_param AND col.`name` NOT IN ('id_users','record_id','user_name','id_actionTriggerTypes','triggerType', 'entry_date', 'user_code');
 
     IF (@sql is null) THEN
         SELECT `name` from view_dataTables where 1=2;
@@ -61,12 +61,13 @@ BEGIN
 			END CASE;
             
             SET @sql = CONCAT('SELECT * FROM (SELECT r.id AS record_id, 
-					r.`timestamp` AS entry_date, r.id_users, u.`name` AS user_name, r.id_actionTriggerTypes, l.lookup_code AS triggerType,', @sql, 
+					r.`timestamp` AS entry_date, r.id_users, u.`name` AS user_name, vc.code AS user_code, r.id_actionTriggerTypes, l.lookup_code AS triggerType,', @sql, 
 					' FROM dataTables t
 					INNER JOIN dataRows r ON (t.id = r.id_dataTables)
 					INNER JOIN dataCells cell ON (cell.id_dataRows = r.id)
 					INNER JOIN dataCols col ON (col.id = cell.id_dataCols)
                     LEFT JOIN users u ON (r.id_users = u.id)
+                    LEFT JOIN validation_codes vc ON (u.id = vc.id_users)
                     LEFT JOIN lookups l ON (l.id = r.id_actionTriggerTypes)
 					WHERE t.id = ', table_id_param, @user_filter, @time_period_filter, @exclude_deleted_filter, 
 					' GROUP BY r.id ) AS r WHERE 1=1  ', filter_param);